% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedMsExperiment.R
\name{LinkedMsExperiment}
\alias{LinkedMsExperiment}
\alias{linkSampleData}
\alias{linkSampleData,MsExperiment-method}
\alias{[,LinkedMsExperiment,ANY,ANY,ANY-method}
\title{Mass spectrometry experiments with explicit links between data entities}
\usage{
linkSampleData(object, ...)

\S4method{linkSampleData}{MsExperiment}(
  object,
  with = character(),
  fromIndex = seq_len(nrow(sampleData(object))),
  toIndex = integer()
)

\S4method{[}{LinkedMsExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\description{
The \code{linkSampleData} function allows to add or define \emph{explicit} links
between samples (i.e. rows in the \code{sampleData} \code{DataFrame}) and other data
elements in the \code{MsExperiment} object. Links can be added between samples
and spectra (in the \verb{@spectra} slot) but also between individual samples and
elements in any other data slot of the \code{MsExperiment}, such as
\code{experimentFiles} or \code{otherData}. Links allow also a n:m mapping between
samples and other data elements.

Importantly, the presence of such links enables a (coherent) subsetting of an
\code{MsExperiment} by samples. Thus, once the link is defined, any subsetting by
sample will also correctly subset the linked data. All other, not linked,
data elements are always retained as in the original \code{MsExperiment}.
}
\section{Accessing data within an \code{LinkedMsExperiment}}{


Naming of data elements in an \code{LinkedExperiment} follows the schema
\code{"<slot name>.<element name>"} (similar to SQL
\code{"<database table>.<column>"}). Using the shortcut
\verb{$<slot name>.<element name>} it is thus also possible to directly access
data elements in the different slots of the object. Note that if a element
with the specified name does not exist in the slot \code{NULL} is returned.
}

\section{Implementation notes}{


Links from samples to any other element are stored as an \code{integer} \code{matrix}
as a list element of the \verb{@sampleDataLinks} slot. The link name (i.e.
the name of the list element) needs to be the name of the slot that is
linked (for slots \verb{@spectra} and \verb{@qfeatures}) or an element within that
slot (for \verb{@experimentFiles}, \verb{@otherData}, \verb{@metadata}). For the latter the
name needs to be in the format \verb{<slot name>.<element name>}, e.g.
\code{"experimentFiles.mzML_files"} to link samples to an element called
\code{"mzML_files"} in \verb{@experimentFiles}.
}

\author{
Johannes Rainer
}
