% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperiment.R
\name{MsExperiment}
\alias{MsExperiment}
\alias{MsExperiment-class}
\alias{metadata,MsExperiment}
\alias{show,MsExperiment-method}
\alias{experimentFiles}
\alias{experimentFiles<-}
\alias{spectra,MsExperiment-method}
\alias{spectra<-}
\alias{metadata,MsExperiment-method}
\alias{metadata<-,MsExperiment-method}
\title{Managing Mass Spectrometry Experiments}
\usage{
MsExperiment()

\S4method{show}{MsExperiment}(object)

experimentFiles(object)

experimentFiles(object) <- value

\S4method{spectra}{MsExperiment}(object)

spectra(object) <- value

\S4method{metadata}{MsExperiment}(x)

\S4method{metadata}{MsExperiment}(x) <- value
}
\arguments{
\item{object}{An instance of class `MsExperiment`}

\item{value}{A `list()` to replace the `MsExperiment`'s metadata.}

\item{x}{An instance of `MsExperiment`.}
}
\description{
The `MsExperiment` class allows the storage, management and processing of all
aspects related to a complete proteomics or metabolomics mass spectrometry
experiment. This includes experimantal design, raw mass spectromtry data as
spectra and chromatograms, quantitative features, and identification data.

For details, see https://rformassspectrometry.github.io/MsExperiment

This package is part of the RforMassSpectrometry initiative:
https://www.rformassspectrometry.org/
}
\details{
- Data files stored in a [MsExperimentFiles()] object.

- Metadata: ...

- Mass spectrometry data, i.e. spectra and their metadata are
  stored as `Spectra()` objects.

- Chromatographic data will be stored as `Chromatograms()` objects.

- Proteomics identification data, i.e peptide-spectrum matches is
  defined as `PSM()` objects. They are generally joined with the
  spectra data.

- Quantification data is stored as `QFeatures()` objects.
}
\examples{

## An empty MsExperiment object
msexp <- MsExperiment()
msexp

example(MsExperimentFiles)
experimentFiles(msexp) <- fls
msexp
}
\author{
Laurent Gatto
}
