% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperiment.R
\docType{class}
\name{MsExperiment}
\alias{MsExperiment}
\alias{MsExperiment-class}
\alias{metadata,MsExperiment}
\alias{show,MsExperiment-method}
\alias{experimentFiles}
\alias{experimentFiles<-}
\alias{sampleData}
\alias{sampleData<-}
\alias{spectra,MsExperiment-method}
\alias{spectra<-}
\alias{metadata,MsExperiment-method}
\alias{metadata<-,MsExperiment-method}
\title{Managing Mass Spectrometry Experiments}
\usage{
MsExperiment()

\S4method{show}{MsExperiment}(object)

experimentFiles(object)

experimentFiles(object) <- value

sampleData(object)

sampleData(object) <- value

\S4method{spectra}{MsExperiment}(object)

spectra(object) <- value

\S4method{metadata}{MsExperiment}(x)

\S4method{metadata}{MsExperiment}(x) <- value
}
\arguments{
\item{object}{An instance of class \code{MsExperiment}}

\item{value}{A \code{list()} to replace the \code{MsExperiment}'s metadata.}

\item{x}{An instance of \code{MsExperiment}.}
}
\description{
The \code{MsExperiment} class allows the storage and management of all
aspects related to a complete proteomics or metabolomics mass
spectrometry experiment. This includes experimantal design, raw
mass spectromtry data as spectra and chromatograms, quantitative
features, and identification data.

For details, see https://rformassspectrometry.github.io/MsExperiment

This package is part of the RforMassSpectrometry initiative:
https://www.rformassspectrometry.org/
}
\details{
An experiment is typically composed of several items
\itemize{
\item Files to data or annotations. There are stored in the
\code{experimentFiles} slot as an instance of class \code{MsExperimentFiles}.
\item General metadata about the experiment, stored as a \code{list} in the
\code{metadata} slot.
\item Mass spectrometry data. Sectra and their metadata are stored as
\code{Spectra()} objects in the \code{spectra} object. Chromatographic is
noy yet supported but will be stored as \code{Chromatograms()} objects
in the \code{chromatorgrams} slot.
\item Quantification data is stored as \code{QFeatures()} objects in the
\code{qfeatures} slot.
\item Any additional data, be it other spectra data, or proteomics
identification data, i.e peptide-spectrum matches is defined as
\code{PSM()} objects can be added as elements to the list stored in
the \code{otherData} slot.
}
}
\section{Slots}{

\describe{
\item{\code{experimentFiles}}{An instance of class \code{MsExperimentFiles} or \code{NULL}.}

\item{\code{spectra}}{An instance of class \code{Spectra} or \code{NULL}.}

\item{\code{qfeatures}}{An instance of class \code{QFeatures} or \code{NULL}.}

\item{\code{otherData}}{A \code{List} to store any additional data objects.}

\item{\code{sampleData}}{A \code{DataFrame} documenting the experimental design.}

\item{\code{metadata}}{A \code{list} to store additional metadata.}
}}

\examples{

## An empty MsExperiment object
msexp <- MsExperiment()
msexp

example(MsExperimentFiles)
experimentFiles(msexp) <- fls
msexp
}
\author{
Laurent Gatto
}
