% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperiment.R
\docType{package}
\name{MsExperiment}
\alias{MsExperiment}
\alias{MsExperiment-package}
\title{Managing Mass Spectrometry Experiments}
\description{
The \code{MsExperiment} class allows the storage, management and processing of all
aspects related to a complete proteomics or metabolomics mass spectrometry
experiment. This includes experimantal design, raw mass spectromtry data as
spectra and chromatograms, quantitative features, and identification data.
}
\details{
For details, see https://rformassspectrometry.github.io/MsExperiment

This package is part of the RforMassSpectrometry initiative:
https://www.rformassspectrometry.org/
}
