% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperiment.R
\docType{class}
\name{MsExperiment}
\alias{MsExperiment}
\alias{MsExperiment-class}
\alias{metadata,MsExperiment}
\alias{show,MsExperiment-method}
\alias{experimentFiles}
\alias{experimentFiles<-}
\alias{sampleData}
\alias{sampleData<-}
\alias{spectra,MsExperiment-method}
\alias{spectra<-}
\alias{metadata,MsExperiment-method}
\alias{metadata<-,MsExperiment-method}
\title{Managing Mass Spectrometry Experiments}
\usage{
MsExperiment()

\S4method{show}{MsExperiment}(object)

experimentFiles(object)

experimentFiles(object) <- value

sampleData(object)

sampleData(object) <- value

\S4method{spectra}{MsExperiment}(object)

spectra(object) <- value

\S4method{metadata}{MsExperiment}(x)

\S4method{metadata}{MsExperiment}(x) <- value
}
\arguments{
\item{object}{An instance of class `MsExperiment`}

\item{value}{A `list()` to replace the `MsExperiment`'s metadata.}

\item{x}{An instance of `MsExperiment`.}
}
\description{
The `MsExperiment` class allows the storage and management of all
aspects related to a complete proteomics or metabolomics mass
spectrometry experiment. This includes experimantal design, raw
mass spectromtry data as spectra and chromatograms, quantitative
features, and identification data.

For details, see https://rformassspectrometry.github.io/MsExperiment

This package is part of the RforMassSpectrometry initiative:
https://www.rformassspectrometry.org/
}
\details{
An experiment is typically composed of several items

- Files to data or annotations. There are stored in the
  `experimentFiles` slot as an instance of class `MsExperimentFiles`.

- General metadata about the experiment, stored as a `list` in the
  `metadata` slot.

- Mass spectrometry data. Sectra and their metadata are stored as
  `Spectra()` objects in the `spectra` object. Chromatographic is
  noy yet supported but will be stored as `Chromatograms()` objects
  in the `chromatorgrams` slot.

- Quantification data is stored as `QFeatures()` objects in the
  `qfeatures` slot.

- Any additional data, be it other spectra data, or proteomics
  identification data, i.e peptide-spectrum matches is defined as
  `PSM()` objects can be added as elements to the list stored in
  the `otherData` slot.
}
\section{Slots}{

\describe{
\item{\code{experimentFiles}}{An instance of class `MsExperimentFiles` or `NULL`.}

\item{\code{spectra}}{An instance of class `Spectra` or `NULL`.}

\item{\code{qfeatures}}{An instance of class `QFeatures` or `NULL`.}

\item{\code{otherData}}{A `List` to store any additional data objects.}

\item{\code{sampleData}}{A `DataFrame` documenting the experimental design.}

\item{\code{metadata}}{A `list` to store additional metadata.}
}}

\examples{

## An empty MsExperiment object
msexp <- MsExperiment()
msexp

example(MsExperimentFiles)
experimentFiles(msexp) <- fls
msexp
}
\author{
Laurent Gatto
}
