% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperiment-functions.R, R/MsExperiment.R
\docType{class}
\name{experimentFiles}
\alias{experimentFiles}
\alias{experimentFiles<-}
\alias{sampleData}
\alias{sampleData<-}
\alias{MsExperiment}
\alias{MsExperiment-class}
\alias{metadata,MsExperiment}
\alias{show,MsExperiment-method}
\alias{spectra,MsExperiment-method}
\alias{spectra<-}
\alias{metadata,MsExperiment-method}
\alias{metadata<-,MsExperiment-method}
\alias{linkSampleData}
\alias{linkSampleData,MsExperiment-method}
\alias{[,MsExperiment,ANY,ANY,ANY-method}
\title{Managing Mass Spectrometry Experiments}
\usage{
experimentFiles(object)

experimentFiles(object) <- value

sampleData(object)

sampleData(object) <- value

MsExperiment()

\S4method{show}{MsExperiment}(object)

\S4method{spectra}{MsExperiment}(object)

spectra(object) <- value

\S4method{metadata}{MsExperiment}(x)

\S4method{metadata}{MsExperiment}(x) <- value

linkSampleData(object, ...)

\S4method{linkSampleData}{MsExperiment}(
  object,
  with = character(),
  sampleIndex = seq_len(nrow(sampleData(object))),
  withIndex = integer(),
  subsetBy = 1L
)

\S4method{[}{MsExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{object}{An instance of class \code{MsExperiment}.}

\item{value}{A \code{list()} to replace the \code{MsExperiment}'s metadata.}

\item{x}{An instance of \code{MsExperiment}.}

\item{...}{optional additional parameters.}

\item{with}{for \code{linkSampleData}: \code{character(1)} defining the data to which
samples should be linked. See section \emph{Linking sample data to other
experimental data} for details.}

\item{sampleIndex}{for \code{linkSampleData}: \code{integer} with the indices of the
samples in \code{sampleData(object)} that should be linked.}

\item{withIndex}{for \code{linkSampleData}: \code{integer} with the indices of the
elements in \code{with} to which the samples (specified by \code{sampleIndex})
should be linked to.}

\item{subsetBy}{for \code{linkSampleData}: optional \code{integer(1)} defining the
dimension on which the subsetting will occurr on the linked data.
Defaults to \code{subsetBy = 1L} thus subsetting will happen on the first
dimension (rows or elements).}

\item{i}{for \code{[}: not supported.}

\item{j}{for \code{[}: an \code{integer}, \code{character} or \code{logical} being the index, the
name (rowname of \code{sampleData}) or a \code{logical} of the samples to subset.}

\item{drop}{for \code{[}: ignored.}
}
\description{
The \code{MsExperiment} class allows the storage and management of all
aspects related to a complete proteomics or metabolomics mass
spectrometry experiment. This includes experimantal design, raw
mass spectromtry data as spectra and chromatograms, quantitative
features, and identification data or any other relevant data files.

For details, see https://rformassspectrometry.github.io/MsExperiment

This package is part of the RforMassSpectrometry initiative:
https://www.rformassspectrometry.org/
}
\section{Slots}{

\describe{
\item{\code{experimentFiles}}{An instance of class \code{MsExperimentFiles} or \code{NULL}.}

\item{\code{spectra}}{An instance of class \code{Spectra} or \code{NULL}.}

\item{\code{qfeatures}}{An instance of class \code{QFeatures} or \code{NULL}.}

\item{\code{otherData}}{A \code{List} to store any additional data objects.}

\item{\code{sampleData}}{A \code{DataFrame} documenting the experimental design.}

\item{\code{metadata}}{A \code{list} to store additional metadata.}
}}

\section{General information}{


An experiment is typically composed of several items
\itemize{
\item Description and information (covariates etc) of each sample from the
experiment. These are stored in the \code{sampleData} slot as a \code{DataFrame},
each row being a sample with columns containing all relevant informatio on
that sample.
\item Files to data or annotations. There are stored in the
\code{experimentFiles} slot as an instance of class \code{MsExperimentFiles}.
\item General metadata about the experiment, stored as a \code{list} in the
\code{metadata} slot.
\item Mass spectrometry data. Sectra and their metadata are stored as
\code{Spectra()} objects in the \code{spectra} object. Chromatographic is
noy yet supported but will be stored as \code{Chromatograms()} objects
in the \code{chromatorgrams} slot.
\item Quantification data is stored as \code{QFeatures()} objects in the
\code{qfeatures} slot.
\item Any additional data, be it other spectra data, or proteomics
identification data, i.e peptide-spectrum matches is defined as
\code{PSM()} objects can be added as elements to the list stored in
the \code{otherData} slot.
}
}

\section{Linking sample data to other experimental data}{


By default, an \code{MsExperiment} is just a loose collection of files and data
related to an experiment, no explicit links or associactions are present
between the samples and related data. Such links can however be created with
the \code{linkSampleData} function. This function can establish links between
individual (or all) samples within the object's \code{sampleData} to individual,
or multiple, data elements or files, such as \code{Spectra} or raw data files.

The presence of such links enables a (coherent) subsetting of an
\code{MsExperiment} by samples. Thus, once the link is defined, any subsetting by
sample will also correctly subset the linked data. All other, not linked,
data elements are always retained as in the original \code{MsExperiment}.

To be able to link different elements within an \code{MsExperiment} it is also
required to \emph{identify} them with a consistent naming scheme. The naming
scheme of slots and data elements within follows an SQL-like scheme, in which
the variable (element) is identified by the name of the database table,
followed by a \code{"."} and the name of the database table column. For
\code{MsExperiment}, the naming scheme is defined as
\code{"<slot name>.<element name>"}. A column called \code{"sample_name"} within the
\code{sampleData} data frame can thus be addressed with
\code{"sampleData.sample_name"}, while \code{spectra.msLevel} would represent the
spectra variable called \code{msLevel} within the \code{Spectra} stored in the
\code{spectra} slot.

Links between sample data rows and any other data element are stored as
\code{integer} matrices within the \code{sampleDataLinks} slot of the object (see also
the vignette for examples and illustrations). Such links can be defined/added
with the \code{linkSampleData} function which adds a relationship between rows in
\code{sampleData} to elements in any other data within the \code{MsExperiment} that
are specified with parameter \code{with}. \code{linkSampleData} supports two different
ways to define the link:
\itemize{
\item Parameter \code{with} defines the data to which the link should be established.
To link samples to raw data files that would for example be available as a
\code{character} in an element called \code{"raw_files"} within the object's
\code{experimentFiles}, \code{with = experimentFiles.raw_files} would have to be
used. Next it is required to specify which samples should be linked with
which elements in \code{with}. This needs to be defined with the parameters
\code{sampleIndex} and \code{withIndex}, both are expected to be \code{integer} vectors
specifying which sample in \code{sampleData} should be linked to which element
in \code{with} (see examples below or vignette for examples and details).
\item As an alternative way, a link could be defined with an SQL-like syntax
that relates a column in \code{sampleData} to a column/element in the data to
which the link should be established. To link for example individual
spectra to the corresponding samples
\code{with = "sampleData.raw_file = spectra.dataOrigin"} could be used assuming
that \code{sampleData} contains a column named \code{"raw_file"} with the (full path)
of the raw data file for each sample from which the spectra were imported.
In this case both \code{sampleIndex} and \code{withIndex} can be omitted, but it is
expected/required that the columns/elements from \code{sampleData} and the data
element to which the link should be established contain matching values.
}

Note that \code{linkSampleData} will \strong{replace} a previously existing link to the
same data element.
}

\section{Subsetting and filtering}{

\itemize{
\item \code{[}: \code{MsExperiment} objects can be subsetted \strong{by samples} with \verb{[, j]}
where \code{j} is the index or a logical defining to which samples the data
should be subsetted. Subsetting by sample will (correctly) subset all
linked data to the respective samples. Not linked data (slots) will be
returned as they are. Subsetting in arbitrary order is supported.
See the vignette for details and examples.
}
}

\examples{

## An empty MsExperiment object
msexp <- MsExperiment()
msexp

example(MsExperimentFiles)
experimentFiles(msexp) <- fls
msexp
}
\author{
Laurent Gatto, Johannes Rainer
}
