% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperimentFiles.R
\docType{class}
\name{MsExperimentFiles}
\alias{MsExperimentFiles}
\alias{MsExperimentFiles-class}
\alias{[,MsExperimentFiles,ANY,ANY,ANY-method}
\alias{[[,MsExperimentFiles,ANY,ANY-method}
\alias{show,MsExperimentFiles-method}
\title{Experiment files}
\usage{
MsExperimentFiles(x)

\S4method{[}{MsExperimentFiles,ANY,ANY,ANY}(x, i)

\S4method{[[}{MsExperimentFiles,ANY,ANY}(x, i)

\S4method{show}{MsExperimentFiles}(object)
}
\arguments{
\item{x}{A named `list()` of characters containing file names.}
}
\description{
The `MsExperimentFiles` class stores files that are part of a mass
spectrometry experiment. The objects are created with the
`MsExperimentFiles()` function that takes a **named** list of
characters as input.
}
\examples{
fls <- MsExperimentFiles(list(mzmls = c("/path/to/f1.mzML", "/path/to/f2.mzML"),
                              mzids = "/another/path/to/id1.mzid"))
fls

## A new MsExperimentFiles containing mzML files
fls[1]

## The actual file names
fls[[1]]
fls[[2]]
}
\author{
Laurent Gatto
}
