% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsExperimentFiles.R
\name{MsExperimentFiles}
\alias{MsExperimentFiles}
\alias{MsExperimentFiles-class}
\title{A class to store experiment files}
\usage{
MsExperimentFiles(..., metadata = list())
}
\arguments{
\item{...}{Either a named list or a set of named vectors. All
elements are coerced to characters.}

\item{metadata}{`list()` holding arbitrary R objects as
annotations.}
}
\description{
The `MsExperimentFiles` class stores files that are part of a mass
spectrometry experiment. The objects are created with the
`MsExperimentFiles()` function.
}
\examples{
fls <- MsExperimentFiles(mzmls = c("/path/to/f1.mzML", "/path/to/f2.mzML"),
                         mzids = "/another/path/to/id1.mzid",
                         fasta = "file.fas")
fls

## A new MsExperimentFiles containing mzML or mzid files
fls[1]
fls["mzids"]

## The actual file names
fls[[1]]
fls[[2]]
fls[["fasta"]]
}
\author{
Laurent Gatto
}
